Add `{% raw %}` and `{% endraw %}` to code blocks
Add `{% raw %}` and `{% endraw %}` to code blocks that contain `{{ ... }}`. In Jekyll-based static site generators, `{{ ... }}` characters are used for Liquid syntax, which causes content to not be displayed, even in code blocks. By enclosing Markdown with `{% raw %}` and `{% endraw %}`, Jekyll will show the text as it should appear, ignoring any possible Liquid syntax.