Migrate docs site from Jekyll to Docusaurus (#235)
* Delete Jekyll file

* Replace Jekyll ignores with Docusaurus ignores

* Add Docusaurus file

* Create _translation-ja-jp.mdx

* Create product name components

* Create unsupported-versions.mdx

* Add Scalar brand asset

* Move file

Move file as part of the migration from Jekyll to Docusaurus.

* Create test-deploy.yml

This workflow checks if Docusaurus will build and deploy properly in production.

* Create deploy.yml

* Delete duplicate file

We no longer add duplicates to manage latest and versioned docs in two separate folders.

* Move file

* Rename file

* Create empty file as a placeholder

* Add image zoom plugin

Add the `docusaurus-plugin-image-zoom` plugin to allow for zooming in on images.

* Remove styles and asterisks from unsupported versions

Keeping styles and asterisks for versions that are no longer supported would require constantly updating the Release Support Policy pages across versions. To avoid this, the styles and asterisks can be removed until we have a better solution.

* Add Mermaid diagram plugin and custom styles

Add the `@docusaurus/theme-mermaid` plugin to support Mermaid diagrams, and add custom styles.

* Delete unnecessary comments

* Remove Jekyll-specific language

* Change links to absolute links

For some reason, using relative links breaks links in this doc. When a visitor clicks a link to a release note, they are taken to the correct release note page. However, if they click the back button to go back to this page and then click a link to a release note, the link is broken because the URL to the release note is incorrect. Changing the relative links to absolute links solves the issue.

* Create release-support-policy.md

* Add styles for cards

* Delete index.md

* Create sidebars.js

* Delete unsupported-versions.md

* Move file

* Delete unnecessary file

* Delete manage-kubernetes.md

Delete this page since it existed only as a stop-gap solution for collapsible side navigation in our previous Jekyll-based docs site. Now that we have collapsible menus, we don't need this page.

* Delete deploy-kubernetes.md

Delete this page since it existed only as a stop-gap solution for collapsible side navigation in our previous Jekyll-based docs site. Now that we have collapsible menus, we don't need this page.

* Change note and tab syntax; change file extension

Change the note and tab syntax from Jekyll/Liquid to Docusaurus/React. Also change the file extension from `.md` to `.mdx` so that we can use React components in Markdown docs.

* Change note syntax

Change the note syntax from Jekyll/Liquid to Docusaurus/React.

* Fix links

* Update .gitignore

* Remove unnecessary redirect

* Update package-lock.json

* Rename folder

* Fix links

* Enclose code in backticks

Building Docusaurus throws an error when `deleteAllById(Iterable<? extends ID> ids)` is not enclosed in backticks. Basically, these APIs should be enclosed in backticks anyway since they're used within code.

* Change file extension from `.md` to `.mdx`

To take advantage of React components in Markdown and for those components to render in Docusaurus, the file extension must be `.mdx`.

* Remove end-of-support banner

This end-of-support banner was specific to Jekyll. Since Docusaurus has a built-in end-of-support (`unmaintained` banner) that can be applied to versions in the Docusaurus configuration file, we'll use that instead.

* Remove unnecessary comment

Commenting out text by using HTML syntax doesn't work with the `.mdx` file extension and aren't really needed within this doc.

* Change comment syntax

HTML syntax doesn't work in docs with the `.mdx` file extension, so I've updated it to the supported syntax.

* Change warning admonition syntax

Change the warning admonition syntax from Jekyll to Docusaurus.

* Update name and version

* Put docs for configuring custom values in category

Put the Helm Charts docs for configuring custom values files in a collapsible category for easier discoverability.

* Change syntax from Jekyll/Liquid to Docusaurus/React

Change syntax for tabs and admonitions from Jekyll/Liquid to Docusaurus/React.

* Add quotation marks to versions

Docusaurus throws an error and cannot build the site when a minor version reaches `X.XX`. In our case, the error displayed is `unknown version (3.1) found`. To get around this, we need to add double quotation marks around version numbers.

* Add links

* Add links

* Comment out i18n setting

Comment out the i18n setting since we don't have any Japanese docs yet.

* Change external links from `.mdx` to `.md`

* Delete `index.mdx`

This doc is awkward as a landing page, so I think removing it would be better. Instead, for now, the first page that the users sees would then be the `overview.mdx`, which is more detailed.

* Delete overview.mdx

* Remove index so visitors are directed to overview

* Move links; re-name home page link

* Re-purpose the home page to show cards

* Add styles for cards

* Add GitHub icon and link

* Change logo link; add GitHub icon and link