Delete `redirect-getting-started.mdx`
This doc was used as a redirect for our Jekyll site. Since the page has very few views, I think it's safe to delete it. If we need to add a redirect for this page int he future for some reason, we'll need to use the redirect plugin in Docusaurus instead.