Revise home page text, remove unnecessary docs, make new category for status code docs (#236)
* Change Card links to lowercase

* Put status codes into category

* Delete `redirect-getting-started.mdx`

This doc was used as a redirect for our Jekyll site. Since the page has very few views, I think it's safe to delete it. If we need to add a redirect for this page int he future for some reason, we'll need to use the redirect plugin in Docusaurus instead.

* Delete unnecessary doc

This doc can be removed since we can now put docs inside collapsible categories in the sidebar navigation.

* Remove comments

* Delete empty line