Add support for tags in docs header (#469)
* Create index.tsx

* Comment out tags in footer

Since we want to have tags at the top of the page we should hide them from the footer.

* Create index.tsx

* Add tags to top of page

Make tags more likely to be seen at first glance on a page by adding components here.

* Add styles for tags at top of page

Add styles to make tags on the same line as `Version` and provide styles to make it look decent in the dark theme.

* Change tag header from `Tags:` to `Editions:`

* Add Font Awesome plugins

* Add styles for Font Awesome question-mark icon

* Add Font Awesome components and question-mark icon

Add Font Awesome components and question-mark icon. The question-mark icon links to our pricing page, which shows features for the available editions.

* Add tooltip for question-mark button/link

* Add tooltip style and border to tags on hover

* Create styles.module.css

* Create index.tsx

* Remove unnecessary styles

These styles are no longer necessary after swizzling the `TagsListInline` component and changing `Tags` to `Editions`.

* Make `Edition` left-aligned

If `Edition` is right aligned, the spacing is awkward if the screen becomes narrow on both desktop and mobile. In such case, at least for now, aligning the `Edition` tags below the version number is less awkward.

* Change component name

Change component name, which was changed in Docusaurus 3.5.0.

* Justify align tags with version

* Comment out "Editions"

Since we'll be showing tags other than editions, we've decided to remove "Editions" from the badge area at the top of the page.

* Add tags

* Add tags

* Remove awkward margin spacing on mobile

* Fix issue with tags being left aligned in Firefox