Add notice about the ScalarDB docs site including both Community and Enterprise docs (#595)
* Create index.tsx

* Comment out tags in footer

Since we want to have tags at the top of the page we should hide them from the footer.

* Create index.tsx

* Add tags to top of page

Make tags more likely to be seen at first glance on a page by adding components here.

* Add styles for tags at top of page

Add styles to make tags on the same line as `Version` and provide styles to make it look decent in the dark theme.

* Change tag header from `Tags:` to `Editions:`

* Add Font Awesome plugins

* Add styles for Font Awesome question-mark icon

* Add Font Awesome components and question-mark icon

Add Font Awesome components and question-mark icon. The question-mark icon links to our pricing page, which shows features for the available editions.

* Add tooltip for question-mark button/link

* Add tooltip style and border to tags on hover

* Create styles.module.css

* Create index.tsx

* Remove unnecessary styles

These styles are no longer necessary after swizzling the `TagsListInline` component and changing `Tags` to `Editions`.

* Make `Edition` left-aligned

If `Edition` is right aligned, the spacing is awkward if the screen becomes narrow on both desktop and mobile. In such case, at least for now, aligning the `Edition` tags below the version number is less awkward.

* Change component name

Change component name, which was changed in Docusaurus 3.5.0.

* Justify align tags with version

* Comment out "Editions"

Since we'll be showing tags other than editions, we've decided to remove "Editions" from the badge area at the top of the page.

* Add tags

* Add tags

* Remove awkward margin spacing on mobile

* Add announcement about unified docs for editions

This announcement is to notify visitors that the docs for both ScalarDB Community and Enterprise editions are available on this docs site. The editions are displayed as tags, which is implemented in #469.

* Fix styling for announcement bar

The close button in the announcement bar was too low and aligned in the middle of the banner.

* Fix issue with tags being left aligned in Firefox

* Fix collapse button background for Firefox

* Revert "Merge branch 'add-support-for-tags-in-doc-header' into announcement/add-notice-about-community-and-enterprise-docs-site"

This reverts commit 71f497816000e8929fbb0c053f2d894debfcc429, reversing
changes made to b14409d5b54b058c6ccd0ff4cbb802846e265aa5.

* Revert "Merge branch 'add-support-for-tags-in-doc-header' into announcement/add-notice-about-community-and-enterprise-docs-site"

This reverts commit ffd99452afeaac4ea0cceb512d07926ea7b5fd33, reversing
changes made to a4ae87444fbe2945a40371407b50451f946c0e12.

* Revise message in announcement banner