Add support dropdown menu to right-hand sidebar of pages (#1027)
* Swizzle component to add support dropdown menu

This component has been swizzled so that we can add a support dropdown menu.

* Add `@typebot.io/react`

* Create AssistantModal.tsx

This is called when the user clicks the AI Assistant button in the Support dropdown menu.

* Create ContactSupportLink.tsx

This is called when the user clicks the Contact Technical Support button in the Support dropdown menu.

* Create SupportDropdownMenu.tsx

* Create SupportDropdownMenu.module.css

* Add Support dropdown menu above TOC

* Make Support button appear with or without ToC

Before this change, the Support button would only appear if the page had a ToC.

* Pass URL as parameter to Typebot

* Show Japanese text on Japanese version of site

* Justify alignment for Support button text and arrow icon

* Make Support button appear static above ToC

By making the Support button static above the ToC, visitors can access the button without having to scroll to the top of the page.

* Hide Support button when hiding ToC on page

Some pages purposely hide the ToC by using `hide_table_of_contents: true`. We don't want the Support button to appear on those pages since the contents of the page are wider (taking up the space where the ToC would be).

* Re-order styles

* Show Support button on mobile

* Make Support button and ToC appear behind language dropdown

When the `zIndex` is too high, the Support button and ToC appear on top of the language dropdown menu. reducing the `zIndex` makes the language dropdown menu appear over the Support button and ToC.

* Revise code comments for clarity

* Fix spelling

* Add support for dark mode

* Delete SupportDropdownMenu.module.css

No longer needed since I added styles to custom.css.

* Add Support button styles

These styles were originally in src/css/SupportDropdownMenu.module.css.

* Change from using `JSX.Element` to `ReactNode`

`ReactNode` should be used since it's what Docusaurus is based on.

* Add GitHub issue link and template

* Fix styles

Styles exist in the custom.css file now rather than a separate CSS file

* Revise wording on links

* Delete ContactSupportLink.tsx

Combined the contents into "src/components/Support/SupportDropdownMenu.tsx" for easier manageability.

* Update SupportDropdownMenu.tsx

Added the contents from "src/components/Support/ContactSupportLink.tsx" for easier manageability.

* Change component link from absolute to relative

Using an absolute link shows as an error for some reason. Changing the absolute link to a relative link doesn't show an error and still imports the component.

* Adjust width and padding for Support button

Need to adjust the width and padding to accommodate changes to the wording.

* Specify Support dropdown arrow size

The dropdown arrow size should be specified here since doing so in the component causes a flash of unstyled content.

* Revised wording in dropdown

* Update font style and link behavior

* Refactor support dropdown to fix errors on build

This refactor is necessary because the error `[cause]: TypeError: Layout_styles_module_default is not a function` occurred. After trying to fix this issue by changing the import for **styles.module.css** to the global **custom.css** file instead, the error `[cause]: ReferenceError: Cannot access 'SupportDropdownMenu' before initialization` occurred. This refactor resolves those issues by ensuring that the support dropdown menu component is initialized first before the Layout component is built.

* Hide Stack Overflow link

We've decided to not show the Stack Overflow link for now. Commented it out so that we can easily show it later.

* Make dropdown menu appear on hover (instead of on click)

Making the dropdown menu appear on hover instead of on click matches the natural behavior of other dropdown menus in Docusaurus.

* Modify support dropdown menu styles

Styles modified to support on-hover behavior instead of on-click behavior and to make the dropdown transitions match the transitions of other dropdown menus in Docusaurus.

* Change dropdown arrow icon

Changed the arrow icon to match the same icon used in other dropdown menus in Docusaurus.

* Fix support dropdown from disappearing

This style keeps the support button dropdown menu open when moving the mouse between the button text and icon.

* Put support button styles in alphabetical order

* Make support button font consistent with other buttons

* Add `some` to English button

Simply `Need help?` results in the arrow being a little too far away from the text. Being a little casual here fixes that.

* Remove comments that were for quick, personal reference

* Add rounded corners to dropdown and on link hover

* Move file

* Move file

* Update import

The component was move in 18027a54a7c19e1c9cf25d29d5bc470173a5c1dc, so this import needs to be updated to reflect that.

* Create AssistantModal.tsx

* Update link to chatbot

* Remove Font Awesome plugin

We decided not to use Font Awesome icons in this component, so we don't need to refer to this plugin here.

* Show AI assistant based on language

* Move back to `Support` folder

* Delete AssistantModal.tsx

This file isn't needed since we can manage the AI assistants in a single file: b160ea86cc9b194732c20da24715a1b371f3043b

* Revert "Update import"

This reverts commit b530b678827cc6e0c3a98595311216e9058824e4.

* Fix GitHub issue link

The GitHub issue link was opening a new tab with the same page that the visitor was on.

* Change dropdown item font color

Change dropdown item font color to match the default color of other fonts in Docusaurus.

* Hide support dropdown menu from home page