Add function to generate log file that contains broken links (#1046)
* Create filter-broken-link-warnings.js

* Make all broken links show warning

A list of all broken links for versions of docs still under Maintenance Support are generated in **brokenLinks.log**. To eliminate potentially redundant entries in the log, only `onBrokenLinks` is set to `warn`.

* Add `brokenLinks.log`

**brokenLinks.log** is for internal/local purposes, so we don't need to include it in the build files.

* Add broken-link log generator to `build` command

* Show broken links for versions under Maintenance Support

To reduce work in fixing broken links for versions that we no longer support, the script checks for broken link for versions of docs that are currently under Maintenance Support.