Add in-site notification feature to Docusaurus (#1168)
* Create NotificationBell.tsx

* Create index.tsx

* Create notifications.js

* Add custom notification function

* Add styles for notification icon and dropdown

* Fix question mark tooltip overlap issue

* Add `@fortawesome/fontawesome-free`

* Fix image background color on zoom

* Allow each language to have its own URL

* Add notification messages

* Make notification message font larger

* Fix broken relative links

Clicking a link in the notification dropdown and then clicking another link resulted in a broken link because the relative path was taken into account when then next link was clicked. For example, if the relative link was `AAA/BBB`, then when the link `CCC/DDD` was selected, the URL would be `AAA/CCC/DDD`. This commit ensures that the base URL is called first and then the relative link in the notification message is applied to that base link.

* Add `Blog post` to notification message

* Add missing ending curly bracket

* Remove duplicate bell icon

* Fix grammar

* Change object name from `languages` to `message`

* Add automatic version tracking for content changes

Implemented content fingerprinting to detect changes in notification messages and URLs. When content changes, the system automatically increments the version number, marking the notification as unread.