Change wording from `New content` to `Recent features` on version home pages; update links (#1376)
* Change UI text from `New content` to `New features`

* Change UI text from `新着コンテンツ` to `新機能`

* Revise code comment

* Make constant name match label name change

* Make style class names match label name change

* Change `new feature` wording to `recent features`

Since the docs shown on the `new features` line won't always be new for that version of docs that the visitor is reading, `recent features` is more appropriate wording here.

* Change `new feature` wording to `recent features`

Since the docs shown on the `new features` line won't always be new for that version of docs that the visitor is reading, `recent features` is more appropriate wording here.

* Change link for diversity's sake

* Revise comment about links in the `Recent features` section

* Change `new feature` wording to `recent features`

Since the docs shown on the `new features` line won't always be new for that version of docs that the visitor is reading, `recent features` is more appropriate wording here.

* Update links

* Add comment about link guidelines

* Add link to doc/feature introduced in 3.15

* Apply suggestions from Gemini review